Add additional notes to report
Add a parser for additional information parsed from a YAML file.
For now, only a note can be specified.

Notes can be specified for any row in the report via a YAML file.
These are read during generation and added in a new column Note.

Additional minor fixes:
* No longer allow unused variables in tsconfig.json
* Exit with error code 1 in case an error occurs

Contributed on behalf of STMicroelectronics

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>