Add onDidChange to theia/plugins
The naming between theia/plugins and vscode/extensions is different.
Therefore the comparator adds the vscode naming
to the theia/plugins module manually.
However, the onDidChange variable was
missing and therefore marked as missing
in the table, while it is available in Theia.
This is fixed with this commit.

Contributed on behalf of STMicroelectronics.