Add missing mapping between extension and plugin
Beforehand `extensionUri`, `extensionKind` and the enum `ExtensionKind`
were flagged as missing on the Theia side.
This is due to the fact, that the names are different.
For the `extensionPath` there was already a mapping in the comparator.
This change also adds them for the two fields and the enum.

Contributed on behalf of STMicroelectronics