Remove `PromiseLike` parsing
After the discussion eclipse-theia/theia/issues#11214
the `PromiseLike` Type was removed from the `theia.d.ts`
file and replaced with `Thenable`.
With this we should stop changing around these values in the
comparator, as this could lead to site effects.
E.g. currently, the `ProviderResult#Thenable` is marked as unsupported.
This is fixed with this.
Furthermore, this will allow us to spot future usages of
`PromiseLike`s more easily.

Follow up for #21.

Contributed on behalf of STMicroelectronics